Renamed the external modules.
